+++ platforms/common/application/menu/index.js
@@ -147,6 +147,8 @@ ready(function() {
         var block = $(last[0].cloneNode(true));
         block.data('mm-id', 'list-' + count);
         block.find('.submenu-items').empty();
+        block.find('[data-mm-base-level]').data('mm-base-level', 1);
+        block.find('.submenu-level').text('Level 1');
         block.after(last);
 
         if (!menumanager.ordering[path]) {